pyproject.toml
[project]
name = "docker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = []

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "pyyaml>=6.0.2",
    "detect-secrets>=1.5.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "types-pyyaml>=6.0.0",
]

[tool.ruff]
target-version = "py313"
line-length = 100
indent-width = 4

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
    "N",
    "W",
    "UP",
    "B",
    "C4",
    "SIM",
    "TCH",
]
ignore = []

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "lf"

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = false
strict_equality = true
show_error_codes = true
show_column_numbers = true
exclude = [
    ".venv",
    ".mypy_cache",
    ".git",
    ".idea",
]
